requirements.txt
transformers
accelerate
datasets
python-chess
stockfish
ctransformers
torch
openai
pandas
nltk
rouge-score
tqdm
bert-score
peft
bitsandbytes
numpy
# Cluster requirements
huggingface_hub
llama-cpp-python
sentence-transformers
